refactor: move app into users
